backend/requirements.txt
fastapi
uvicorn[standard]
openai
pinecone-client
pydantic
